refactor: 파일명 변경 어떤 컴포넌트에 대한 스켈레톤인지 명시 Skeleton > MovieListSkeleton
